backend/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
django-restframework==0.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
Markdown==3.4.1
pytz==2022.2.1
sqlparse==0.4.2
flake8==4.0.1
lazy-object-proxy==1.4.1
pylint-django==2.5.3
virtualenv==20.10.0
python-dotenv==0.19.2
trycourier
